+++ src/xrGame/game_sv_item_respawner.cpp
@@ -103,7 +103,7 @@ void item_respawn_manager::check_to_spawn(CSE_Abstract* item)
     }
 }*/
 
-CSE_Abstract* item_respawn_manager::make_respawn_entity(shared_str const& section_name, u8 addons, u16 count_of_ammo)
+CSE_Abstract* item_respawn_manager::make_respawn_entity(shared_str const& section_name, u32 addons, u16 count_of_ammo)
 {
     R_ASSERT(m_server);
     R_ASSERT(m_server->GetGameState());
